refactor: 경매 관련 페이지 shared 폴더 관련 파일 분리 및 tanstackquery 내부 api 함수 분리 타입 분리
