+++ apps/web/shared/hooks/useRealTimeBid.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 
-import { supabase } from '../shared/lib/supabase/supabase';
 import { Bid } from '../types/db';
+import { supabase } from '../lib/supabase/supabase';
 
 export function useRealtimeBids(auctionId: string | undefined, onNewBid: (bid: Bid) => void) {
   const channelRef = useRef<ReturnType<typeof supabase.channel> | null>(null);
