+++ apps/web/shared/types/auction/index.ts
@@ -1,0 +1,21 @@
+export interface AuctionBidParams {
+  auctionId: string;
+  bidderId: string;
+  bidPrice: number;
+}
+
+export interface AuctionBidResponse {
+  success?: boolean;
+  error?: string;
+}
+
+export interface CreateAuctionPayload {
+  name: string;
+  category: string;
+  description: string;
+  start_price: number;
+  start_time: string;
+  end_time: string;
+  thumbnail: string;
+  images: string[];
+}
